make it work better
- simplify internal directory structure
- move `cat.png` into `rich-message` module
- update `.travis.yml` to target node 4 & 5, remove outdated optimizations
- add usage example
