+++ .travis.yml
@@ -1,9 +1,7 @@
 language: node_js
 node_js:
-  - 'iojs'
-sudo: false
+  - '4'
+  - '5'
 cache:
   directories:
     - node_modules
-script:
-  - npm test
