+++ CHANGELOG.md
@@ -1,13 +1,20 @@
 # rich-message change log
 
 All notable changes to this project will be documented in this file.
+
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/moose-team/rich-message/compare/v1.0.1...HEAD)
+
 * fix channel name parsing bug ([#8](https://github.com/moose-team/rich-message/issues/8))
 * add tests ([#9](https://github.com/moose-team/rich-message/pull/9))
+* simplify internal directory structure
+* move `cat.png` into `rich-message` module
+* update `.travis.yml` to target node 4 & 5, remove outdated optimizations
+* add usage example
 
 ## [1.0.1](https://github.com/moose-team/rich-message/compare/v1.0.0...v1.0.1)
+
 * fix channel link formatting ([#6](https://github.com/moose-team/rich-message/pull/6))
 * use MOOSE Team MIT License ([#7](https://github.com/moose-team/rich-message/pull/7))
