+++ README.md
@@ -1,5 +1,7 @@
 # rich-message
 
+> Turn a plain message into a rich HTML message
+
 [![npm][npm-image]][npm-url]
 [![travis][travis-image]][travis-url]
 
@@ -7,8 +9,6 @@
 [npm-url]: https://www.npmjs.com/package/rich-message
 [travis-image]: https://img.shields.io/travis/moose-team/rich-message.svg?style=flat-square
 [travis-url]: https://travis-ci.org/moose-team/rich-message
-
-
 
 ## Install
 
@@ -20,7 +20,24 @@ npm install rich-message
 
 ```js
 var richMessage = require('rich-message')
+var basicMessage = {
+  text: 'hi maxogden', // text entered by a user
+  username: 'mafintosh', // github username
+  timestamp: Date.now()
+}
+var username = 'maxogden' // current user's github username (used for highlighting)
+
+var output = richMessage(basicMessage, username)
+// { text: 'hi maxogden',
+//   username: 'mafintosh',
+//   timestamp: 1458939703123,
+//   anon: false,
+//   avatar: 'https://github.com/mafintosh.png',
+//   timeago: '2:01 PM',
+//   html: '<div class="highlight"><p>hi maxogden</p><p></p></div>' }
 ```
+
+This module is currently very tightly coupled with [friends](https://github.com/moose-team/friends) (sorry). If you'd like to help make it more usable for other projects, please fork and contribute!
 
 ## Contributing
